Extract logic out of service, into FusedLocationProviderServiceImpl
