+++ lost/src/test/java/com/mapzen/android/lost/internal/FusedLocationProviderServiceTest.java
@@ -1,0 +1,5 @@
+package com.mapzen.android.lost.internal;
+
+public class FusedLocationProviderServiceTest {
+
+}
